backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-genai==1.1.0
httpx==0.28.1
pydantic==2.10.4
pyyaml==6.0.2
aiosqlite==0.20.0
websockets==14.1
pytest==8.3.4
pytest-asyncio==0.25.0
